pyproject.toml
[project]
name = "git-helper"
dynamic = ["version"]
description = "A tool for managing git commits with structured workflows"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Rajeev Karamcheti", email = "rkaramc@gmail.com"}
]
dependencies = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "gitpython>=3.1.0",
    "readchar>=4.0.0"
]

[dependency-groups]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0"
]

[project.scripts]
gw-commit = "git_helper.main:cli"

[project.urls]
Homepage = "https://github.com/rkaramc/git-workflow-tool"
Repository = "https://github.com/rkaramc/git-workflow-tool.git"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/git_helper"]

[tool.hatch.version]
path = "src/git_helper/__init__.py"
